pe: simplify generate_hash()
Copying the value of datasize_in to two further variables and then using
all three randomly in the code makes it hard to read.

datasize_in is never changed in generate_hash() so we can do with this
parameter alone. Let's declare it as const to verify that its value
is invariant. Rename it to datasize.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>